// src/shared/router/types/router.type.ts
import { ReactNode } from 'react'

export type RouteType = {
  active?: boolean
  children?: RouteType[]
  component: ReactNode
  exact?: boolean
  icon?: string
  nameMenu: string
  path: string
  title?: string
  type?: 'modal' | 'page'
}

export type RoutesGroupType = {
  title: string
  routes: RouteType[]
}

export type RouterType = {
  routesGroups: RoutesGroupType[]
}

export default RouterType